+++ src/base_dumper/common/queries.py
@@ -12,20 +12,48 @@ from sqlglot.errors import (
 )
 
 
+DIALECT = {
+    "athena": "athena",
+    "bigquery": "bigquery",
+    "clickhouse": "clickhouse",
+    "databricks": "databricks",
+    "denodo": "denodo",
+    "doris": "doris",
+    "drill": "drill",
+    "duckdb": "duckdb",
+    "greenplum": "postgres",
+    "hive": "hive",
+    "materialize": "materialize",
+    "mysql": "mysql",
+    "oracle": "oracle",
+    "postgres": "postgres",
+    "presto": "presto",
+    "redshift": "redshift",
+    "risingwave": "risingwave",
+    "snowflake": "snowflake",
+    "spark": "spark",
+    "spark2": "spark2",
+    "sqlite": "sqlite",
+    "sqlserver": "tsql",
+    "starrocks": "starrocks",
+    "tableau": "tableau",
+    "teradata": "teradata",
+    "trino": "trino",
+}
 EXECUTE_PATTERN = compile(r"^(with|select|show|grant|describe)\s", IGNORECASE)
 FIRST_WORD_PATTERN = compile(r"^\s*([a-zA-Z]+)")
 STRIP_CHARS = "; \t\n\r"
 UNKNOWN = "Unknown"
 
 
-def get_query_kind(query: str) -> str:
+def get_query_kind(query: str, dialect: str = "postgres") -> str:
     """Get kind of query."""
 
     if not query:
         return UNKNOWN
 
     try:
-        ast = parse_one(query, error_level=ErrorLevel.IGNORE)
+        ast = parse_one(query, read=dialect, error_level=ErrorLevel.IGNORE)
 
         if not ast:
             match = FIRST_WORD_PATTERN.match(query)
@@ -39,14 +67,17 @@ def get_query_kind(query: str) -> str:
         return UNKNOWN
 
 
-def query_formatter(queries: str) -> str:
+def query_formatter(queries: str, dialect: str = "postgres") -> str:
     """Reformat query, removing comments."""
 
-    statements = transpile(queries, comments=False)
+    statements = transpile(queries, read=dialect, comments=False)
     return ";\n".join(statements).strip(STRIP_CHARS)
 
 
-def chunk_query(queries: str | None) -> tuple[list[str], list[str]]:
+def chunk_query(
+    queries: str | None,
+    dialect: str = "postgres",
+) -> tuple[list[str], list[str]]:
     """Chunk multiquery to queries."""
 
     if not queries:
@@ -54,7 +85,7 @@ def chunk_query(queries: str | None) -> tuple[list[str], list[str]]:
 
     all_queries = [
         part.strip(STRIP_CHARS)
-        for part in transpile(queries, comments=False)
+        for part in transpile(queries, read=dialect, comments=False)
         if part.strip(STRIP_CHARS)
     ]